chunk_store: widen cs_nblocks to uint64_t
cs_nblocks was uint32_t but chunk_store_write and chunk_store_lookup
accept uint64_t offsets.  The grow function capped at 1M entries so
this was safe in practice, but the type mismatch meant offsets above
UINT32_MAX could silently bypass bounds checks after a truncating
cast.

+++ lib/nfs4/server/chunk_store.c
@@ -81,7 +81,7 @@ static int chunk_store_grow(struct chunk_store *cs, uint64_t new_cap)
 	memcpy(nb, cs->cs_blocks, cs->cs_nblocks * sizeof(*nb));
 	free(cs->cs_blocks);
 	cs->cs_blocks = nb;
-	cs->cs_nblocks = (uint32_t)cap;
+	cs->cs_nblocks = cap;
 	return 0;
 }
